Delete perl-style matching delimiters.
grep does not use pattern matching delimiters since its parameter is directly the
regexp pattern, so the original pattern included a literal forward slash before
the patterns for the supported VCS directories.

As etckeeper cd's into the /etc, the command to find possible hardlinks passes
the relative path of the git objects beginning with ".git/" to the VCS filter
that did not match because of the leading slash in the pattern.