store all permissions of files and directories, even those with standard permissions of 644 and 755
This is unfortunately necessary in order to support etckeeper init on a
checkout that was made with a nonstandard umask, in which case the files
that were expected to be 644 and 755, won't be. Closes: #649701

Thanks to Дмитрий Матросов for reporting the bug and developing a fixup
script (attached to the bug) which could be used if you've already
encountered this problem.