Omit files ignored by git from .etckeeper file
Add processing of the .gitignore file into filter_ignore() function.

The .gitignore content is a list of globs and not regular expressions, so it is
not possible to use same approach darcs uses. Instead, we use git commands to
generate list of files that exist and need to be excluded as pattern file.