Avoid propigating nonzero exit status of any command called by etckeeper out to the caller. This is to avoid making apt give up because git is misconfigured and cannot commit, or similar. Closes: #760011
I absoltely do not want etckeeper to dive into the rats nest of making git
be able to commit reliably. git commit has about a dozen failure modes, and
it's up to the admin to avoid breaking their system by not giving it a
hostname, or writing a broken ~/.gitconfig, or not giving uid 0 a user
name, etc.

I considered making only apt.conf's etckeeper calls be guarded with ||
true, but this would also need to be done for the other package managers
that can call etckeeper. And it's better to let etckeeper
indicate which script failed, rather than seeing an error message and apt
continuing without an indication of why git commit was being called then.